extract: checkIsDisabledStep from AddTemplateContext
Continuing with extraction of logic out of the top level context.
This time extracting a check for disabled steps.
A new hook useCheckIsDisabledStep was created for the extracted logic.
